Spike, how to add tile layer on top of map in Strava route builder

+++ LTC2.Desktopclients.WindowsClient/Forms/MainForm.cs
@@ -232,6 +232,13 @@ namespace LTC2.Desktopclients.WindowsClient.Forms
 
             foreach (var cookie in stravaCookies)
             {
+                var cookieName = cookie.Name;
+
+                if (_appSettings.SkipCookiesWhileDeleting != null && _appSettings.SkipCookiesWhileDeleting.Contains(cookieName))
+                {
+                    continue;
+                }
+
                 webView.CoreWebView2.CookieManager.DeleteCookie(cookie);
             }
 
@@ -239,6 +246,13 @@ namespace LTC2.Desktopclients.WindowsClient.Forms
 
             foreach (var cookie in rwgpsCookies)
             {
+                var cookieName = cookie.Name;
+
+                if (_appSettings.SkipCookiesWhileDeleting != null && _appSettings.SkipCookiesWhileDeleting.Contains(cookieName))
+                {
+                    continue;
+                }
+
                 webView.CoreWebView2.CookieManager.DeleteCookie(cookie);
             }
         }
@@ -263,15 +277,19 @@ namespace LTC2.Desktopclients.WindowsClient.Forms
 
             if (_refreshEnabledFor != null)
             {
+                var found = false;
+
                 foreach (var url in _refreshEnabledFor)
                 {
                     if (webView.CoreWebView2.Source.StartsWith(url))
                     {
-                        btnRefresh.Enabled = true;
+                        found = true;
 
                         break;
                     }
                 }
+
+                btnRefresh.Enabled = found;
 
                 if (btnRefresh.Enabled)
                 {
@@ -280,6 +298,87 @@ namespace LTC2.Desktopclients.WindowsClient.Forms
                 else
                 {
                     btnRefresh.BackgroundImage = Properties.Resources.refresh3;
+                }
+
+                if (webView.CoreWebView2.Source.StartsWith("https://www.strava.com/maps/create"))
+                {
+                    var script = @"
+                        function GetMapbox() {
+                            console.log('looking for mapbox element');
+                        
+                            var elements = document.getElementsByClassName('mapboxgl-map');
+                            if (elements[0] == null) {
+                                console.log('mapbox element not found');
+                            } else {
+                                var element = elements[0];
+                                Object.entries(element).find(([k, _]) => k.startsWith('__react'))[1].return.memoizedProps.mapboxRef((x) => (window.routeMap = x, x));
+            
+                                if (window.routeMap != null) {
+                                    console.log('route map found');
+
+                                    console.log('try adding vector layer');
+
+                                    window.routeMap.on('style.load', () => {
+                                        console.log('changed layer style re-add layer to: ' + window.routeMap.getStyle().sprite);
+
+                                        AddTileLayer();
+                                    });
+
+                                    AddTileLayer();
+            
+                                    return '1';
+                                } 
+                            
+                                console.log('route map is null try again');
+                            }
+
+                            return '0';
+                        }
+
+                        function AddTileLayer() {
+                            if (window.routeMap != null) {
+                                console.log('adding vector layer');
+                                window.routeMap.addSource('ltc2tiles', {
+                                    'type': 'vector',
+                                    'tiles': [' http://localhost:50000/api/Tiles/tile/{z}/{x}/{y}.pbf'],
+                                    'minzoom': 3,
+                                    'maxzoom': 17
+                                });
+                                window.routeMap.addLayer({
+                                    'id': 'ltc2tiles',
+                                    'type': 'line',
+                                    'source': 'ltc2tiles',
+                                    'source-layer': 'geojsonLayer',
+                                    'layout': {
+                                        'line-cap': 'round',
+                                        'line-join': 'round'
+                                    },
+                                    'paint': {
+                                        'line-opacity': 1.0,
+                                        'line-color': 'rgb(0, 0, 0)',
+                                        'line-width': 2
+                                    },
+                                    slot: 'middle'
+                                });
+                            } else {
+                                console.log('route map is null cannot add tile layer');
+                            }
+                        }
+
+                        GetMapbox();
+                    ";
+
+                    Task.Run(async () =>
+                    {
+                        var r = await webView.ExecuteScriptAsync(script);
+
+                        while (r == null || r.Trim('"') != "1")
+                        {
+                            await Task.Delay(500);
+
+                            r = await webView.ExecuteScriptAsync(script);
+                        }
+                    });
                 }
             }
         }
